Fixing bugs on multiprocessing

+++ Tools for networks/version 2.1/Tools for Networkers_v2.1.py
@@ -19,6 +19,8 @@ import datetime
 import io
 from contextlib import redirect_stdout #This is to redirect stdout
 from multiprocessing import Process #This is for multi-processing
+#import multiprocessing
+#https://docs.python.org/2/library/multiprocessing.html#multiprocessing.Process.daemon
 import os
 
 import telnetlib         # This mainly used for TAB-1 such as telnet feature. 
@@ -35,6 +37,11 @@ from scapy.all import *  # This mainly used for TAB-2 such as packet building se
     
 
 def _sending_packets(cap_name,count,interval):    
+    #name = multiprocessing.current_process().name
+    #id = multiprocessing.current_process().pid
+    #print(name, id)
+    #print(os.getpid())
+
     if interval == 0:
         sendp(cap_name,loop=1)
     else :
@@ -73,14 +80,14 @@ class Form(QtWidgets.QMainWindow):
         self.lineEdit_6.setText("1")#In TAB-1, "2). Time sleep (sec) Interleaving between commands (default 1sec)"
         self.dateTimeEdit # In TAB-1, "> Beginning"
         self.dateTimeEdit_2# In TAB-1, "> End"
-        self._thread_1 = None #Initialization to the un-defined thread value Used for 'Repetitive Tasks'
-        self._tl_1 = None     #Initialization to the un-defined Telnet object value Used for 'Repetitive Tasks'
+        self._thread_1 = None #Initialization to the un-defined thread value Used at 'Repetitive Tasks'
+        self._tl_1 = None     #Initialization to the un-defined Telnet object value Used at 'Repetitive Tasks'
 
      #Initialization for TAB-2
         self.lineEdit_7.setText("1") # In TAB-2, "3). The number of packets to send"
         self.lineEdit_8.setText("1") # In TAB-2, "4). PPS"
         self.current_datatime = QDateTime.currentDateTime()
-             
+        self.p_list = [] # This is for multiprocess to take in the processes running as to list. Later This list used to process join()/terminate()    
 
     def _savefiledialog(self): # _savefiledialog at memu bar
         try:
@@ -279,9 +286,10 @@ class Form(QtWidgets.QMainWindow):
     @QtCore.pyqtSlot() # This slot is for 'Running' in TAB-2
     def slot_8st(self): 
         try:
-            self.p_list = [] # This is for multiprocess to take in the processes running as to list. Later This list used to process join()/terminate()
-            p = Process(target=_sending_packets,args=(self.new_modified_packets, int(self.lineEdit_7.text()), int(self.lineEdit_8.text()))) # multiprocessing, and passing the arguments.
+          
+            p = Process(target=_sending_packets, name= "process name", args=(self.new_modified_packets, int(self.lineEdit_7.text()), int(self.lineEdit_8.text()))) # multiprocessing, and passing the arguments.
             p.start()  
+         #   print(p.pid)
             self.p_list.append(p)        
         except:
             self.textBrowser_2.setText("No packet to send. please check if you miss the sequence, such as 'running checksum or else' or not proper filename, else ")
@@ -291,10 +299,15 @@ class Form(QtWidgets.QMainWindow):
 #Used for 'Packet Builder'
     @QtCore.pyqtSlot() # This slot is for 'Stopping' in TAB-2
     def slot_9st(self):        
-        for p in self.p_list: # This is to terminate all multi-processes
-            p.terminate()   
-        for p in self.p_list: # This is to join() all multi-processes during termation.
-            p.join() 
+        if self.p_list != None:
+            for p in self.p_list: # This is to terminate/join all multi-processes
+                print(p.pid)
+                p.terminate()  
+            for p in self.p_list: # This is to terminate/join all multi-processes
+                print(p.pid)
+                p.join()  
+            self.p_list = [] # This is to inistialize the list
+
